backend/requirements.txt
Django==5.0
djangorestframework==3.15.0
djoser==2.2.0
django-filter==24.3
Pillow==10.0.0
PyYAML==6.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9